Moved empire name arrays to separate file
